move host and mgmt-net nodes to appropriate files
